Clarify skill audit support paths
