+++ server/src/services/portable-path.ts
@@ -1,0 +1,12 @@
+export function normalizePortablePath(input: string) {
+  const parts: string[] = [];
+  for (const segment of input.replace(/\\/g, "/").replace(/^\.\/+/, "").replace(/^\/+/, "").split("/")) {
+    if (!segment || segment === ".") continue;
+    if (segment === "..") {
+      if (parts.length > 0) parts.pop();
+      continue;
+    }
+    parts.push(segment);
+  }
+  return parts.join("/");
+}
